front_inst_jmp.ccにJMP系命令を移動
